+++ src/raytracing/raytracing.c
@@ -6,7 +6,7 @@
 /*   By: aalbrech <aalbrech@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 12:41:59 by aalbrech          #+#    #+#             */
-/*   Updated: 2025/04/29 14:21:36 by aalbrech         ###   ########.fr       */
+/*   Updated: 2025/04/29 21:36:30 by aalbrech         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -128,15 +128,15 @@ of the pixel at coordinate [pixel_x, pixel_y], in 3d space.
 */
 static t_xyz	get_ray_direction(int pixel_x, int pixel_y, t_camera *camera)
 {
-	float	scalar_x;
-	float	scalar_y;
+	double	scalar_x;
+	double	scalar_y;
 	t_xyz	direction;
-	float	fisheye_radius;
-	float	phi_angle;
+	double	fisheye_radius;
+	double	phi_angle;
 
-	scalar_x = (2 * (((float)pixel_x + 0.5f) / IMG_WIDTH) - 1)
+	scalar_x = (2 * (((double)pixel_x + 0.5f) / IMG_WIDTH) - 1)
 		* camera->aspect_ratio * camera->fov_scale;
-	scalar_y = (1 - 2 * (((float)pixel_y + 0.5f) / IMG_HEIGHT))
+	scalar_y = (1 - 2 * (((double)pixel_y + 0.5f) / IMG_HEIGHT))
 		* camera->fov_scale;
 	fisheye_radius = sqrt(scalar_x * scalar_x + scalar_y * scalar_y);
 	fisheye_radius = fisheye_radius * 0.1;
@@ -214,8 +214,8 @@ static void	set_detailed_camera(t_camera *camera)
 	else
 		camera->right_view = vec_normalize(
 				vec_cross(camera->world_up, camera->orientation));
-	camera->aspect_ratio = (float)IMG_WIDTH / (float)IMG_HEIGHT;
-	camera->fov_scale = tanf(((float)camera->fov * M_PI / 180.0) / 2.0);
+	camera->aspect_ratio = (double)IMG_WIDTH / (double)IMG_HEIGHT;
+	camera->fov_scale = tan(((double)camera->fov * M_PI / 180.0) / 2.0);
 }
 
 /*
